Revert "#32694: Migrate op to new infra: layernorm_pre_all_gather_op (#33526)"
This reverts commit 48920b57935a38f70546bd7b53b2f5bb2c9b955c.
