+++ ttnn/cpp/ttnn/operations/normalization/CMakeLists.txt
@@ -31,7 +31,6 @@ target_sources(
             layernorm/layernorm.hpp
             layernorm/device/layernorm_types.hpp
             layernorm_distributed/device/layernorm_distributed_types.hpp
-            layernorm_distributed/device/layernorm_pre_all_gather_device_operation.hpp
             rmsnorm/rmsnorm.hpp
             softmax/device/softmax_operation_types.hpp
             softmax/device/softmax_device_operation.hpp
@@ -61,10 +60,9 @@ target_sources(
         layernorm/device/layernorm_op_multi_core_sharded.cpp
         layernorm/layernorm.cpp
         layernorm_distributed/device/layernorm_post_all_gather_op.cpp
-        layernorm_distributed/device/layernorm_pre_all_gather_device_operation.cpp
-        layernorm_distributed/device/layernorm_pre_all_gather_program_factory.cpp
-        layernorm_distributed/device/layernorm_pre_all_gather_2d_program_factory.cpp
+        layernorm_distributed/device/layernorm_pre_all_gather_op.cpp
         layernorm_distributed/device/multi_core/layernorm_post_all_gather_op_multi_core.cpp
+        layernorm_distributed/device/multi_core/layernorm_pre_all_gather_op_multi_core.cpp
         layernorm_distributed/layernorm_post_all_gather.cpp
         layernorm_distributed/layernorm_pre_all_gather.cpp
         rmsnorm/rmsnorm.cpp
