+++ ttnn/cpp/ttnn/operations/normalization/layernorm_distributed/device/layernorm_pre_all_gather_op.hpp
@@ -1,0 +1,38 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <optional>
+
+#include "ttnn/operations/core/compute_kernel/compute_kernel_config.hpp"
+#include "ttnn/tensor/tensor.hpp"
+#include "ttnn/operations/core/core.hpp"
+
+#include "layernorm_distributed_types.hpp"
+
+namespace ttnn::operations::normalization {
+
+tt::tt_metal::operation::ProgramWithCallbacks layernorm_pre_allgather_multi_core(
+    const Tensor& a,
+    Tensor& output,
+    LayerNormDistributedType norm_type,
+    DeviceComputeKernelConfig compute_kernel_config,
+    std::optional<bool> use_2d_core_grid = std::nullopt,
+    LayerNormDistributedDefaultProgramConfig program_config = LayerNormDistributedDefaultProgramConfig{});
+
+struct LayerNormPreAllGather {
+    LayerNormDistributedType norm_type;
+    const DataType dtype;
+    const DeviceComputeKernelConfig compute_kernel_config;
+    std::optional<bool> use_2d_core_grid;
+    LayerNormDistributedDefaultProgramConfig program_config;
+
+    void validate(const std::vector<Tensor>& input_tensors) const;
+    std::vector<TensorSpec> compute_output_specs(const std::vector<Tensor>& input_tensors) const;
+    tt::tt_metal::operation::ProgramWithCallbacks create_program(
+        const std::vector<Tensor>& input_tensors, std::vector<Tensor>& output_tensors) const;
+};
+
+}  // namespace ttnn::operations::normalization
