+++ ttnn/cpp/ttnn/operations/normalization/layernorm_distributed/layernorm_pre_all_gather.cpp
@@ -4,9 +4,10 @@
 
 #include "layernorm_pre_all_gather.hpp"
 
-#include "device/layernorm_pre_all_gather_device_operation.hpp"
+#include "device/layernorm_pre_all_gather_op.hpp"
 #include "ttnn/operations/normalization/layernorm/device/layernorm_device_operation.hpp"
 #include "ttnn/device.hpp"
+namespace operation = tt::tt_metal::operation;
 
 namespace ttnn::operations::normalization {
 
@@ -36,14 +37,15 @@ ttnn::Tensor ExecuteLayerNormPreAllGather::invoke(
             LayerNormType::LAYERNORM,
             DistributedLayerNormStage::PRE_ALL_GATHER);
     } else {
-        return ttnn::prim::layernorm_pre_all_gather(
-            input_tensor,
-            LayerNormDistributedType::LAYERNORM,
-            dtype,
-            kernel_config_val,
-            std::nullopt,
-            distributed_program_config,
-            std::nullopt);
+        return operation::run(
+                   LayerNormPreAllGather{
+                       .norm_type = LayerNormDistributedType::LAYERNORM,
+                       .dtype = dtype,
+                       .compute_kernel_config = kernel_config_val,
+                       .use_2d_core_grid = std::nullopt,
+                       .program_config = distributed_program_config},
+                   {input_tensor})
+            .at(0);
     }
 }
